pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "costest"
version = "0.1.0"
description = "Cost Estimate Generator"
readme = "README.md"
authors = [{ name = "Road Tools Automation" }]
requires-python = ">=3.9"
dependencies = [
    "numpy==1.26.4",
    "pandas==1.5.3",
    "openpyxl==3.1.2",
    "python-dotenv>=1.0.0,<2.0.0",
    "xlrd>=2.0.1,<3.0.0",
    "openai>=1.0.0,<2.0.0",
    "reportlab>=4.0.0,<5.0.0",
    "PyPDF2==3.0.1",
]

[tool.setuptools.packages.find]
where = ["src"]
